+++ src/macro.cpp
@@ -282,20 +282,20 @@ bool Macro::shouldStep() {
     if (g.stepFrame) return true;
     if (Global::getCurrentFrame() == 0) return true;
 
-    if (g.ignoreFrame != -1) return true;
-    if (g.ignoreJumpButton != -1) return true;
+    // if (g.ignoreFrame != -1) return true;
+    // if (g.ignoreJumpButton != -1) return true;
 
-    if (g.delayedFrameReleaseMain[0] != -1) return true;
-    if (g.delayedFrameReleaseMain[1] != -1) return true;
+    // if (g.delayedFrameReleaseMain[0] != -1) return true;
+    // if (g.delayedFrameReleaseMain[1] != -1) return true;
 
-    if (g.delayedFrameInput[0] != -1) return true;
-    if (g.delayedFrameInput[1] != -1) return true;
+    // if (g.delayedFrameInput[0] != -1) return true;
+    // if (g.delayedFrameInput[1] != -1) return true;
 
-    for (int x = 0; x < 2; x++) {
-        for (int y = 0; y < 2; y++) {
-            if (g.delayedFrameRelease[x][y] != -1) return true;
-        }
-    }
+    // for (int x = 0; x < 2; x++) {
+    //     for (int y = 0; y < 2; y++) {
+    //         if (g.delayedFrameRelease[x][y] != -1) return true;
+    //     }
+    // }
 
     return false;
 }